Revert "Add slash commands and other various changes from devel (#27)"
This reverts commit 059a124e9483f5df2d739dc5e9301a961b3cbf0c.

+++ src/main/java/yt/graven/gravensupport/commands/ticket/create/interactions/FirstSentenceHandler.java
@@ -1,11 +1,10 @@
 package yt.graven.gravensupport.commands.ticket.create.interactions;
 
-import java.awt.*;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import net.dv8tion.jda.api.EmbedBuilder;
-import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
-import net.dv8tion.jda.api.events.interaction.component.StringSelectInteractionEvent;
+import net.dv8tion.jda.api.MessageBuilder;
+import net.dv8tion.jda.api.entities.TextChannel;
+import net.dv8tion.jda.api.events.interaction.component.SelectMenuInteractionEvent;
 import net.dv8tion.jda.api.utils.MiscUtil;
 import org.springframework.stereotype.Component;
 import yt.graven.gravensupport.commands.ticket.Ticket;
@@ -14,15 +13,18 @@ import yt.graven.gravensupport.utils.interactions.IIInteractionAction;
 import yt.graven.gravensupport.utils.messages.Embeds;
 import yt.graven.gravensupport.utils.messages.TMessage;
 
+import java.awt.*;
+import java.util.Optional;
+
 @Component
 @RequiredArgsConstructor
-public class FirstSentenceHandler implements IIInteractionAction<StringSelectInteractionEvent> {
+public class FirstSentenceHandler implements IIInteractionAction<SelectMenuInteractionEvent> {
 
     private final TicketManager ticketManager;
     private final Embeds embeds;
 
     @Override
-    public void run(StringSelectInteractionEvent event) {
+    public void run(SelectMenuInteractionEvent event) {
         Optional<Ticket> ticket = ticketManager.get(MiscUtil.parseLong(((TextChannel) event.getChannel()).getTopic()));
         if (ticket.isEmpty()) {
             event.deferReply(true)
@@ -36,12 +38,11 @@ public class FirstSentenceHandler implements IIInteractionAction<StringSelectInt
             return;
         }
 
-        String content =
-                switch (event.getInteraction().getSelectedOptions().get(0).getValue()) {
-                    case "bonsoir" -> ":wave: Bonsoir, comment pouvons-nous vous aider ?";
-                    case "bonjour" -> ":wave: Bonjour, comment pouvons-nous vous aider ?";
-                    default -> "";
-                };
+        String content = switch (event.getInteraction().getSelectedOptions().get(0).getValue()) {
+            case "bonsoir" -> ":wave: Bonsoir, comment pouvons-nous vous aider ?";
+            case "bonjour" -> ":wave: Bonjour, comment pouvons-nous vous aider ?";
+            default -> "";
+        };
 
         if (content.isEmpty()) return;
 
@@ -51,20 +52,21 @@ public class FirstSentenceHandler implements IIInteractionAction<StringSelectInt
                 .setFooter("")
                 .setColor(Color.GREEN);
 
-        TMessage.create()
-                .setContent(content)
-                .sendMessage(ticket.get().getFrom())
-                .queue(
-                        (msg) -> {
-                            event.deferEdit().queue();
-                            TMessage.from(embed.build())
-                                    .sendMessage(event.getChannel())
-                                    .queue();
-                            event.getMessage().delete().queue();
-                        },
-                        (err) -> {
-                            event.reply(embeds.errorMessage(err.getMessage()).build())
-                                    .queue();
-                        });
+        ticket.get().confirmSendToUser(new MessageBuilder().setContent(content).build()).thenAccept((msg) -> {
+
+            event.deferEdit().queue();
+
+            TMessage.from(embed.build())
+                    .sendMessage(event.getChannel())
+                    .queue();
+
+            event.getMessage().delete().queue();
+
+        }).exceptionally((error) -> {
+            event
+                    .reply(embeds.errorMessage(error.getMessage()).build())
+                    .queue();
+            return null;
+        });
     }
 }
